Move all event related stuff to @colony/events
