--- packages/event-metadata/CHANGELOG.md
@@ -1,24 +1,0 @@
-# @colony/event-metadata
-
-## 3.0.1-next.1
-
-### Patch Changes
-
-- 50eefa91: Improve compatibility between ESM package.json generation and changesets:
-  - Use template file for package.json stub files (for changeset compatibility)
-  - Add .js extensions also for type imports and exports
-  - Add type entries to all package.json files
-  - Rename all .js config files to .cjs
-
-## 3.0.1-next.0
-
-### Patch Changes
-
-- 8de9d9ca: Upgrade TypeDoc to v0.24.8 and adjust link syntax style
-- 542ffad8: Adjust import/export statements to be compatible with ESM only within node.js
-
-## 3.0.0
-
-### Major Changes
-
-- Complete rewrite
