--- packages/event-metadata/tsconfig.build.json
@@ -1,18 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": [
-    "src/index.ts"
-  ],
-  "exclude": [
-    "**/__tests__"
-  ],
-  "compilerOptions": {
-    "noEmit": false,
-    "declaration": false,
-    "plugins": [
-      {
-        "transform": "typia/lib/transform"
-      }
-    ]
-  }
-}
