--- packages/event-metadata/tsconfig.json
@@ -1,27 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "inlineSourceMap": false,
-    "lib": [
-      "esnext"
-    ],
-    "listEmittedFiles": false,
-    "listFiles": false,
-    "moduleResolution": "nodenext",
-    "noEmit": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": true,
-    "pretty": true,
-    "resolveJsonModule": true,
-    "rootDir": "./src",
-    "skipLibCheck": true,
-    "strictNullChecks": true,
-    "strict": true,
-    "traceResolution": false
-  },
-  "include": [
-    "src"
-  ]
-}
