+++ packages/events/src/ColonyEventManager.ts
@@ -3,11 +3,16 @@ import type { BlockTag } from '@ethersproject/abstract-provider';
 
 import { constants, providers, EventFilter, BaseContract } from 'ethers';
 import { addressesAreEqual, SignerOrProvider } from '@colony/core';
-import { MetadataType, MetadataTypeMap } from '@colony/event-metadata';
 
-import type { Ethers6Filter } from '../types.js';
-import { getLogs, nonNullable } from '../utils.js';
-import { IpfsAdapter, IpfsMetadata, MetadataEvent } from '../ipfs/index.js';
+import {
+  type IpfsAdapter,
+  type MetadataEvent,
+  type MetadataTypeMap,
+  IpfsMetadata,
+  MetadataType,
+} from './ipfs/index.js';
+import type { Ethers6Filter } from './types.js';
+import { getLogs, nonNullable } from './utils.js';
 
 /**
  * A valid eventsource (currently just an ethers.js {@link BaseContract})
