+++ packages/events/src/index.ts
@@ -1,48 +1,13 @@
-export * from './contracts/index.js';
+export {
+  ColonyEvent,
+  ColonyEventManager,
+  ColonyEventManagerOptions,
+  ColonyTopic,
+  ColonyFilter,
+  ColonyMultiFilter,
+  type EventSource,
+} from './ColonyEventManager.js';
 
-// We need to manually export those as they are ambiguous as they exist in multiple event contracts
-export type {
-  MetaTransactionExecutedEvent,
-  MetaTransactionExecutedEventFilter,
-  MetaTransactionExecutedEventObject,
-  RecoveryModeEnteredEvent,
-  RecoveryRoleSetEvent,
-  RecoveryModeExitedEventObject,
-  RecoveryRoleSetEventFilter,
-  RecoveryStorageSlotSetEvent,
-  RecoveryRoleSetEventObject,
-  RecoveryStorageSlotSetEventFilter,
-  RecoveryModeExitedEventFilter,
-  RecoveryModeExitApprovedEventObject,
-  RecoveryModeEnteredEventFilter,
-  RecoveryModeExitedEvent,
-  RecoveryModeEnteredEventObject,
-  RecoveryModeExitApprovedEventFilter,
-  RecoveryModeExitApprovedEvent,
-  RecoveryStorageSlotSetEventObject,
-  ReputationMinerPenalisedEvent,
-  ReputationMinerPenalisedEventFilter,
-  ReputationMinerPenalisedEventObject,
-} from './contracts/IColonyNetworkEvents.js';
+export * from './ipfs/index.js';
 
-// All ColonyExtensions have those so we export them from the OG contract
-export type {
-  ExtensionInitialisedEvent,
-  LogSetAuthorityEventObject,
-  LogSetOwnerEvent,
-  LogSetOwnerEventFilter,
-  LogSetAuthorityEventFilter,
-  ExtensionInitialisedEventObject,
-  LogSetAuthorityEvent,
-  ExtensionInitialisedEventFilter,
-  LogSetOwnerEventObject,
-} from './contracts/ColonyExtensionEvents.js';
-
-export * from './contracts/CoinMachineEvents.js';
-export * from './contracts/IColonyEvents.js';
-export * from './contracts/IColonyNetworkEvents.js';
-export * from './contracts/MetaTxTokenEvents.js';
-export * from './contracts/OneTxPaymentEvents.js';
-export * from './contracts/TokenLockingEvents.js';
-export * from './contracts/VotingReputationEvents.js';
-export * from './contracts/WhitelistEvents.js';
+export * from './types.js';
