+++ packages/events/src/ipfs/index.ts
@@ -8,3 +8,7 @@ export { default as PinataAdapter } from './PinataAdapter.js';
 export { default as CloudflareReadonlyAdapter } from './CloudflareReadonlyAdapter.js';
 
 export type { default as IpfsAdapter } from './IpfsAdapter.js';
+
+export * from './metadata.js';
+export * from './constants.js';
+export * from './types.js';
