+++ packages/sdk/package.json
@@ -49,7 +49,7 @@
   },
   "dependencies": {
     "@colony/core": "^1.2.0-next.2",
-    "@colony/event-metadata": "^3.0.1-next.1",
+    "@colony/events": "^0.2.1-next.1",
     "@colony/tokens": "^0.1.3-next.1",
     "@urql/core": "^2.5.0",
     "cross-fetch": "^3.1.5",
@@ -60,7 +60,6 @@
   },
   "devDependencies": {
     "@colony/contractor": "^1.1.0-next.2",
-    "@colony/events": "^0.2.1-next.1",
     "@ethersproject/abi": "^5.7.0",
     "@ethersproject/abstract-provider": "^5.7.0",
     "@ethersproject/abstract-signer": "^5.7.0",
