+++ packages/sdk/src/ColonyNetwork/tokens/ERC20Token.ts
@@ -1,5 +1,8 @@
 import type { BigNumber, BigNumberish } from 'ethers';
-import type { ApprovalEventObject, TransferEventObject } from '@colony/events';
+import type {
+  ApprovalEventObject,
+  TransferEventObject,
+} from '@colony/events/types';
 
 import {
   ERC20Token as ERC20TokenType,
