--- packages/sdk/src/events/index.ts
@@ -1,9 +1,0 @@
-export {
-  ColonyEvent,
-  ColonyEventManager,
-  ColonyEventManagerOptions,
-  ColonyTopic,
-  ColonyFilter,
-  ColonyMultiFilter,
-  type EventSource,
-} from './ColonyEventManager.js';
