+++ packages/sdk/src/utils.ts
@@ -1,14 +1,8 @@
-import { ContractReceipt } from 'ethers';
-
-import type { Log } from '@ethersproject/abstract-provider';
-import type { JsonRpcProvider } from '@ethersproject/providers';
 import type { Interface } from '@ethersproject/abi';
 
-import {
-  Ethers6Filter,
-  Ethers6FilterByBlockHash,
-  ParsedLogTransactionReceipt,
-} from './types.js';
+import { ContractReceipt } from 'ethers';
+
+import { ParsedLogTransactionReceipt } from './types.js';
 
 /** Extract event args from a contract receipt */
 // TODO: Make it possible to retrieve multiple events of the same kind
@@ -56,21 +50,4 @@ export const extractCustomEvent = <T>(
     }
   }
   return undefined;
-};
-
-/** Version of `getLogs` that also supports filtering for multiple addresses */
-export const getLogs = async (
-  filter:
-    | Ethers6Filter
-    | Ethers6FilterByBlockHash
-    | Promise<Ethers6Filter | Ethers6FilterByBlockHash>,
-  provider: JsonRpcProvider,
-): Promise<Log[]> => {
-  const usedFilter = await filter;
-  return provider.send('eth_getLogs', [usedFilter]);
-};
-
-/** Use this to filter empty undefinied values from arrays in a type-safe way */
-export const nonNullable = <T>(value: T): value is NonNullable<T> => {
-  return value !== null && value !== undefined;
 };
